rdkit/requirements.txt
# Function dependencies, for example:
# package>=version
rdkit-pypi
uamc-qed

